Clean up Demucs
