+++ react-ui/src/pages/demucs.tsx
@@ -1,56 +1,45 @@
 import React, { useState } from "react";
 import { Template } from "../components/Template";
-import FileInput, { getAudioURL } from "../components/FileInput";
-import { AudioPlayer } from "../components/MemoizedWaveSurferPlayer";
-import { WaveSurferOptions } from "wavesurfer.js";
 import useLocalStorage from "../hooks/useLocalStorage";
+import { GradioFile } from "./api/demucs_musicgen";
+import { useRouter } from "next/router";
+import { AudioInput, AudioOutput } from "../components/AudioComponents";
 
-type AudioOutput = {
-  name: string;
-  data: string;
-  size?: number;
-  is_file?: boolean;
-  orig_name?: string;
+type TypedGradioFile = GradioFile & {
   type_name?: string;
 };
 
-function addTypeNameToAudioOutput(audioOutput: AudioOutput, typeName: string) {
-  return { ...audioOutput, type_name: typeName };
-}
+const typeNames = ["drums", "bass", "other", "vocals"];
 
-function addTypeNamesToAudioOutputs(
-  audioOutputs: AudioOutput[],
+const addTypeNamesToAudioOutputs = (
+  audioOutputs: TypedGradioFile[],
   typeNames: string[]
-) {
-  return audioOutputs.map((audioOutput, index) =>
-    addTypeNameToAudioOutput(audioOutput, typeNames[index])
-  );
-}
+) =>
+  audioOutputs.map((audioOutput, index) => ({
+    ...audioOutput,
+    type_name: typeNames[index],
+  }));
 
 type DemucsParams = {
-  file: string | null;
+  file?: string;
 };
-const initialState = {
+
+const initialState: DemucsParams = {
   file: "https://www.mfiles.co.uk/mp3-downloads/gs-cd-track2.mp3",
 };
 
 const DemucsPage = () => {
-  // const [data, setData] = useState<AudioOutput[] | null>(null);
-  // local storage
-  const [data, setData] = useLocalStorage<AudioOutput[] | null>("data", null);
-
-  // const [demucsParams, setDemucsParams] = useState<DemucsParams>({
-  //   // file: null,
-  //   file: "https://www.mfiles.co.uk/mp3-downloads/gs-cd-track2.mp3",
-  // });
-  // local storage
+  const [data, setData] = useLocalStorage<TypedGradioFile[] | null>(
+    "data",
+    null
+  );
   const [demucsParams, setDemucsParams] = useLocalStorage<DemucsParams>(
     "demucs-tab",
     initialState
   );
 
   async function demucs() {
-    const response = await fetch("/api/demucs_demucs", {
+    const response = await fetch("/api/demucs_musicgen", {
       method: "POST",
       body: JSON.stringify(demucsParams),
     });
@@ -59,34 +48,35 @@ const DemucsPage = () => {
     setData(result?.data);
   }
 
-  const typeNames = ["drums", "bass", "other", "vocals"];
   const sampleWithTypeNames =
     data && addTypeNamesToAudioOutputs(data, typeNames);
+
+  const router = useRouter();
+
+  const sendToDemucs = (file: string) => {
+    setDemucsParams(
+      (demucsParams) => ({ ...demucsParams, file } as DemucsParams)
+    );
+    router.push("/demucs");
+  };
+
+  const sendAudioTo = (audio: string | undefined) => {
+    if (audio) sendToDemucs(audio);
+  };
 
   return (
     <Template>
       <div className="p-4">
         <div>
-          <MusicFileInput
+          <AudioInput
             url={demucsParams?.file}
-            callback={(melody) => {
+            callback={(file) => {
               setDemucsParams({
                 ...demucsParams,
-                file: melody,
+                file,
               });
             }}
-            sendAudioTo={(audio) => {
-              setData([
-                {
-                  name: "demucs",
-                  data: audio || "",
-                  type_name: "melody",
-                  is_file: false,
-                  orig_name: "melody",
-                  size: 0,
-                },
-              ]);
-            }}
+            sendAudioTo={sendAudioTo}
           />
 
           <button
@@ -100,19 +90,11 @@ const DemucsPage = () => {
               (item) => item.type_name === typeName
             );
             return (
-              <MusicFileOutput
+              <AudioOutput
                 key={typeName}
                 audioOutput={audioOutput}
                 label={typeName}
-                sendAudioTo={(audio) => {
-                  setData([
-                    {
-                      name: "demucs",
-                      data: audio || "",
-                      type_name: typeName,
-                    },
-                  ]);
-                }}
+                sendAudioTo={sendAudioTo}
               />
             );
           })}
@@ -123,115 +105,3 @@ const DemucsPage = () => {
 };
 
 export default DemucsPage;
-
-const MusicFileInput = ({
-  callback,
-  sendAudioTo,
-  url,
-}: {
-  callback: (melody: string | null) => void;
-  sendAudioTo: (audio: string | undefined) => void;
-  url: string | undefined;
-}) => (
-  <div className="mt-4 border border-gray-300 p-2 rounded">
-    <label className="text-sm">Input file:</label>
-    <FileInput
-      callback={(file: File | undefined) => {
-        const melody = getAudioURL(file) || null;
-        callback(melody);
-      }}
-    />
-    <AudioPlayerHelper url={url} sendAudioTo={sendAudioTo} />
-  </div>
-);
-
-const MusicFileOutput = ({
-  audioOutput,
-  label,
-  sendAudioTo,
-}: {
-  audioOutput?: AudioOutput;
-  label: string;
-  sendAudioTo: (audio: string | undefined) => void;
-}) => {
-  return (
-    <div className="mt-4 border border-gray-300 p-2 rounded">
-      <p className="text-sm">{label}</p>
-      {audioOutput && (
-        <>
-          <AudioPlayerHelper url={audioOutput.data} sendAudioTo={sendAudioTo} />
-          <AudioOutputInfo audioOutput={audioOutput} />
-        </>
-      )}
-    </div>
-  );
-};
-
-const AudioPlayerHelper = (
-  props: Omit<WaveSurferOptions, "container"> & {
-    volume?: number;
-    sendAudioTo: (audio: string | undefined) => void;
-  }
-) => {
-  return (
-    <>
-      <AudioPlayer
-        height={100}
-        waveColor="#ffa500"
-        progressColor="#d59520"
-        // volume={0.4}
-        barWidth={2}
-        barGap={1}
-        barRadius={2}
-        // {...props}
-        volume={props.volume || 0.4}
-        url={props.url}
-      />
-      <button
-        className="bg-gray-500 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded"
-        onClick={() => props.sendAudioTo(props.url)}
-      >
-        Send Audio
-      </button>
-      <a
-        // className="bg-gray-500 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded"
-        // equalize with <button>
-        className="bg-gray-500 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded inline-block"
-        href={props.url}
-        download
-      >
-        Download
-      </a>
-    </>
-  );
-};
-
-const AudioOutputInfo = ({ audioOutput }: { audioOutput: AudioOutput }) => {
-  return (
-    <div className="border border-gray-300 p-2 rounded">
-      <p className="text-sm">Info</p>
-      <div className="flex flex-col text-xs">
-        <div className="flex flex-row">
-          <div className="w-1/3">Name</div>
-          <div className="w-1/3">{audioOutput.name}</div>
-        </div>
-        <div className="flex flex-row">
-          <div className="w-1/3">Original name</div>
-          <div className="w-1/3">{audioOutput.orig_name}</div>
-        </div>
-        <div className="flex flex-row">
-          <div className="w-1/3">Type name</div>
-          <div className="w-1/3">{audioOutput.type_name}</div>
-        </div>
-        <div className="flex flex-row">
-          <div className="w-1/3">Size</div>
-          <div className="w-1/3">{audioOutput.size}</div>
-        </div>
-        <div className="flex flex-row">
-          <div className="w-1/3">Is file</div>
-          <div className="w-1/3">{audioOutput.is_file ? "true" : "false"}</div>
-        </div>
-      </div>
-    </div>
-  );
-};
